pyproject.toml
[tool.poetry]
name = "energy_demand_australia"
version = "0.1.0"
description = "Energy demand and resulting retail price development in the Australian states."
authors = ["danielvogler <davogler@ethz.ch>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
pandas = "^1.4.2"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
